move plots to their own file
