+++ neurodsp/spectral.py
@@ -2,7 +2,6 @@
 
 import numpy as np
 from scipy import signal
-import matplotlib.pylab as plt
 
 ###################################################################################################
 ###################################################################################################
@@ -350,37 +349,6 @@ def spectral_hist(x, Fs, window='hann', nperseg=None, noverlap=None,
     spect_hist = np.flipud(spect_hist)
 
     return freq, power_bins, spect_hist
-
-
-def plot_spectral_hist(freq, power_bins, spect_hist, psd_freq=None, psd=None):
-    """Plot the spectral histogram.
-
-    Parameters
-    ----------
-    freq : array_like, 1d
-        Frequencies over which the histogram is calculated.
-    power_bins : array_like, 1d
-        Power bins within which histogram is aggregated.
-    spect_hist : ndarray, 2d
-        Spectral histogram to be plotted.
-    psd_freq : array_like, 1d, optional
-        Frequency axis of the PSD to be plotted.
-    psd : array_like, 1d, optional
-        PSD to be plotted over the histograms.
-    """
-
-    # automatically scale figure height based on number of bins
-    plt.figure(figsize=(8, 12 * len(power_bins) / len(freq)))
-    # plot histogram intensity as image and automatically adjust aspect ratio
-    plt.imshow(spect_hist, extent=[freq[0], freq[-1], power_bins[0], power_bins[-1]], aspect='auto')
-    plt.xlabel('Frequency (Hz)', fontsize=15)
-    plt.ylabel('Log10 Power', fontsize=15)
-    plt.colorbar(label='Probability')
-
-    if psd is not None:
-        # if a PSD is provided, plot over the histogram data
-        plt.plot(psd_freq[np.logical_and(psd_freq >= freq[0], psd_freq <= freq[-1])], np.log10(
-            psd[np.logical_and(psd_freq >= freq[0], psd_freq <= freq[-1])]), color='w', alpha=0.8)
 
 
 def morlet_transform(x, f0s, Fs, w=7, s=.5):
